pyproject.toml
[project]
name = "django-react2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django==5.1.5",
    "django-vite==3.0.6",
    "inertia-django==1.1.0",
]
